compliance-tracker/backened/requirements.txt
flask
requests
bs4
google-generativeai
